Refactory: SecurityContextInterceptor preHandle() 메소드 추상화
-
