+++ src/main/java/nextstep/auth/authorization/SecurityContextInterceptor.java
@@ -5,6 +5,7 @@ import javax.servlet.http.HttpServletResponse;
 
 import org.springframework.web.servlet.HandlerInterceptor;
 
+import nextstep.auth.context.SecurityContext;
 import nextstep.auth.context.SecurityContextHolder;
 
 public abstract class SecurityContextInterceptor implements HandlerInterceptor {
@@ -12,4 +13,19 @@ public abstract class SecurityContextInterceptor implements HandlerInterceptor {
 	public void afterCompletion(HttpServletRequest request, HttpServletResponse response, Object handler, Exception ex) {
 		SecurityContextHolder.clearContext();
 	}
+
+	@Override
+	public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
+		if (SecurityContextHolder.getContext().getAuthentication() != null) {
+			return true;
+		}
+
+		SecurityContext securityContext = getSecurityContext(request);
+		if (securityContext != null) {
+			SecurityContextHolder.setContext(securityContext);
+		}
+		return true;
+	}
+
+	public abstract SecurityContext getSecurityContext(HttpServletRequest request);
 }
