+++ src/main/java/nextstep/auth/authorization/TokenSecurityContextPersistenceInterceptor.java
@@ -3,14 +3,12 @@ package nextstep.auth.authorization;
 import java.util.Map;
 
 import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import nextstep.auth.context.Authentication;
 import nextstep.auth.context.SecurityContext;
-import nextstep.auth.context.SecurityContextHolder;
 import nextstep.auth.token.JwtTokenProvider;
 
 public class TokenSecurityContextPersistenceInterceptor extends SecurityContextInterceptor {
@@ -21,21 +19,14 @@ public class TokenSecurityContextPersistenceInterceptor extends SecurityContextI
     }
 
     @Override
-    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
-        if (SecurityContextHolder.getContext().getAuthentication() != null) {
-            return true;
-        }
-
+    public SecurityContext getSecurityContext(HttpServletRequest request) {
         String credentials = AuthorizationExtractor.extract(request, AuthorizationType.BEARER);
+
         if (!jwtTokenProvider.validateToken(credentials)) {
-            return true;
+            return null;
         }
 
-        SecurityContext securityContext = extractSecurityContext(credentials);
-        if (securityContext != null) {
-            SecurityContextHolder.setContext(securityContext);
-        }
-        return true;
+        return extractSecurityContext(credentials);
     }
 
     private SecurityContext extractSecurityContext(String credentials) {
